move files-dependency-builder to teambit.dependencies
