+++ scopes/dependencies/dependencies/files-dependency-builder/build-tree.spec.ts
