+++ scopes/dependencies/dependencies/files-dependency-builder/build-tree.ts
