+++ scopes/dependencies/dependencies/files-dependency-builder/dependency-tree/Config.ts
