+++ scopes/dependencies/dependencies/files-dependency-builder/dependency-tree/index.spec.ts
