+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-css/index.ts
