+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-es6/index.ts
