+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-less/index.ts
