+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-sass/index.ts
