+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-scss/index.ts
