+++ scopes/dependencies/dependencies/files-dependency-builder/detectives/detective-typescript/index.ts
