+++ scopes/dependencies/dependencies/files-dependency-builder/detector-hook.ts
