+++ scopes/dependencies/dependencies/files-dependency-builder/filing-cabinet/index.ts
