+++ scopes/dependencies/dependencies/files-dependency-builder/generate-tree-madge.ts
