+++ scopes/dependencies/dependencies/files-dependency-builder/index.ts
