+++ scopes/dependencies/dependencies/files-dependency-builder/missing-handler.ts
