+++ scopes/dependencies/dependencies/files-dependency-builder/path-map.ts
@@ -8,7 +8,7 @@
 import R from 'ramda';
 
 import { processPath } from './generate-tree-madge';
-import { Specifier } from './types/dependency-tree-type';
+import { Specifier } from '@teambit/legacy/dist/consumer/component/dependencies/dependency';
 
 export type PathMapDependency = {
   importSource: string; // dependency path as it has been received from dependency-tree lib
