+++ scopes/dependencies/dependencies/files-dependency-builder/precinct/index.ts
