+++ scopes/dependencies/dependencies/files-dependency-builder/types/index.ts
