+++ scopes/dependencies/dependencies/index.ts
@@ -1,5 +1,6 @@
 import { DependenciesAspect } from './dependencies.aspect';
 
 export type { DependenciesMain } from './dependencies.main.runtime';
+export { DetectorHook, DependencyDetector, FileContext } from './files-dependency-builder/detector-hook';
 export default DependenciesAspect;
 export { DependenciesAspect };
