+++ scopes/dependencies/dependency-resolver/dependency-detector.ts
@@ -1,5 +1,1 @@
-export {
-  DetectorHook,
-  DependencyDetector,
-  FileContext,
-} from '@teambit/legacy/dist/consumer/component/dependencies/files-dependency-builder/detector-hook';
+export { DetectorHook, DependencyDetector, FileContext } from '@teambit/dependencies';
