+++ scopes/dependencies/dependency-resolver/dependency-resolver.main.runtime.ts
@@ -24,7 +24,7 @@ import {
 } from '@teambit/legacy/dist/constants';
 import { ExtensionDataList } from '@teambit/legacy/dist/consumer/config/extension-data';
 import { componentIdToPackageName } from '@teambit/pkg.modules.component-package-name';
-import { DetectorHook } from '@teambit/legacy/dist/consumer/component/dependencies/files-dependency-builder/detector-hook';
+import { DetectorHook } from '@teambit/dependencies';
 import { Http, ProxyConfig, NetworkConfig } from '@teambit/legacy/dist/scope/network/http';
 import { onTagIdTransformer } from '@teambit/snapping';
 import LegacyComponent from '@teambit/legacy/dist/consumer/component';
